--- arbitrator/espresso-crypto-helper/Cargo.toml
@@ -1,52 +1,0 @@
-
-[package]
-name = "espresso-crypto-helper"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-ark-bn254 = "0.4"
-ark-ed-on-bn254 = "0.4"
-ark-ff = "0.4"
-ark-serialize = { version = "0.4", features = ["derive"] }
-ark-std = { version = "0.4", default-features = false }
-base64 = "0.22"
-bytesize = "1.3"
-derive_more = "0.99.17"
-digest = "0.10"
-ethers-core = "2.0.14"
-sha2 = "0.10"
-base64-bytes = "0.1"
-paste = "1.0"
-serde_json = "1.0"
-committable = {version = "0.2", features = ["ark-serialize"]}
-ethereum-types = "0.14.1"
-tagged-base64 = "0.4"
-lazy_static = "1.4"
-serde = { version = "1.0.195", features = ["derive"] }
-itertools = "0.12"
-either =  {version = "1.11.0", features = ["serde"]}
-
-jf-crhf = { version = "0.1.0", git = "https://github.com/EspressoSystems/jellyfish", tag = "0.4.5" }
-jf-utils = { git = "https://github.com/EspressoSystems/jellyfish", tag = "0.4.5", features = [], default-features = false}
-jf-vid = { version = "0.1.0", git = "https://github.com/EspressoSystems/jellyfish", tag = "0.4.5", features = [
-  "std",
-], default-features = false}
-jf-merkle-tree = { version = "0.1.0", git = "https://github.com/EspressoSystems/jellyfish", tag = "0.4.5", features = [
-  "std",
-], default-features = false}
-jf-pcs = { version = "0.1.0", git = "https://github.com/EspressoSystems/jellyfish", tag = "0.4.5", features = [
-  "std",
-], default-features = false }
-jf-rescue = { version = "0.1.0", git = "https://github.com/EspressoSystems/jellyfish", tag = "0.4.5", features = [
-  "std",
-], default-features = false}
-trait-set = "0.3.0"
-typenum = { version = "1.15.0", default-features = false, features = [
-    "no_std",
-] }
-num-traits = "0.2.17"
-derivative = "2.2"
-
-[dev-dependencies]
-ark-srs = { version = "0.3.1" }
