--- arbitrator/espresso-crypto-helper/src/full_payload.rs
@@ -1,8 +1,0 @@
-mod ns_proof;
-mod ns_table;
-mod payload;
-
-pub use ns_proof::NsProof;
-pub use ns_table::{NsIndex, NsTable};
-
-pub use payload::PayloadByteLen;
