--- arbitrator/espresso-crypto-helper/src/namespace_payload.rs
@@ -1,8 +1,0 @@
-mod iter;
-mod ns_payload;
-mod ns_payload_range;
-mod tx_proof;
-mod types;
-
-pub use ns_payload::NsPayloadOwned;
-pub use ns_payload_range::NsPayloadRange;
