--- arbitrator/espresso-crypto-helper/vid_srs.json
@@ -1,1 +1,0 @@
-[5,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,249,230,49,207,189,34,111,38,253,239,131,186,97,251,231,171,98,78,190,123,168,209,107,255,67,249,159,40,40,6,54,45,170,10,112,5,212,123,240,59,100,176,247,78,203,61,211,248,240,162,236,254,231,204,82,130,244,195,99,229,121,42,185,166,219,137,228,207,21,151,128,212,130,59,138,190,25,73,54,189,135,81,200,124,137,67,158,163,166,72,205,60,228,214,10,2,67,69,158,23,152,197,51,106,108,140,24,44,111,20,97,192,10,96,210,7,20,82,156,180,146,61,234,226,186,158,167,161,118,140,215,125,138,226,57,182,7,222,71,91,232,103,156,194,54,190,135,208,45,8,92,144,114,46,122,185,203,237,11,33,101,226,121,255,196,55,110,254,210,246,234,12,28,192,114,52,238,47,69,188,187,4,105,47,171,163,58,184,49,228,209,152,214,96,132,233,17,76,177,73,144,41,127,136,58,178,129,238,30,130,27,98,71,103,194,185,19,190,176,246,65,206,86,42,124,44,239,166,174,161,44,209,192,86,134,199,56,99,125,130,213,222,37,231,228,109,143,125,84,218,221,227,45,143,169,154,237,246,146,217,92,189,222,70,221,218,94,247,212,34,67,103,121,68,92,94,102,0,106,66,118,30,31,18,239,222,0,24,194,18,243,174,183,133,228,151,18,231,169,53,51,73,170,241,37,93,251,49,183,191,96,114,58,72,13,146,147,147,142,25,170,125,250,102,1,204,230,76,123,211,67,12,105,231,209,227,143,64,203,141,128,113,171,74,235,109,140,219,165,94,200,18,91,151,34,209,220,218,172,85,243,142,179,112,51,49,75,188,149,51,12,105,173,153,158,236,117,240,95,88,208,137,6,9,176,131,136,147,236,31,35,126,139,7,50,59,7,68,89,159,78,151,181,152,179,181,137,188,194,188,55,184,213,196,24,1,193,131,147,192,250,48,254,78,139,3,142,53,122,216,81,234,232,222,145,7,88,78,255,231,199,241,246,81,178,1,14,38,85,94,204,218,212,135,74,133,162,206,230,150,63,221,230,17,94,97,229,20,66,91,71,86,42,99,192,192,163,189,254,34,228,95,106,218,128,60,65,238,164,155,249,65,70,160,242,156,133,114,154,187,193,86,81,210,227,15,17,247,105,99,252,4,2,0,0,0,0,0,0,0,237,246,146,217,92,189,222,70,221,218,94,247,212,34,67,103,121,68,92,94,102,0,106,66,118,30,31,18,239,222,0,24,194,18,243,174,183,133,228,151,18,231,169,53,51,73,170,241,37,93,251,49,183,191,96,114,58,72,13,146,147,147,142,25,170,125,250,102,1,204,230,76,123,211,67,12,105,231,209,227,143,64,203,141,128,113,171,74,235,109,140,219,165,94,200,18,91,151,34,209,220,218,172,85,243,142,179,112,51,49,75,188,149,51,12,105,173,153,158,236,117,240,95,88,208,137,6,9,176,131,136,147,236,31,35,126,139,7,50,59,7,68,89,159,78,151,181,152,179,181,137,188,194,188,55,184,213,196,24,1,193,131,147,192,250,48,254,78,139,3,142,53,122,216,81,234,232,222,145,7,88,78,255,231,199,241,246,81,178,1,14,38,85,94,204,218,212,135,74,133,162,206,230,150,63,221,230,17,94,97,229,20,66,91,71,86,42,99,192,192,163,189,254,34,228,95,106,218,128,60,65,238,164,155,249,65,70,160,242,156,133,114,154,187,193,86,81,210,227,15,17,247,105,99,252,4]
